Automating AWS operations with Boto API
This project’s functionality needs to be launched via main.py that serves as an entry point to the program.

The option (1) selected in the main leads to launching login.py that validates the inputted credentials against the content of the passwords.txt file. When option (2) is selected, a user can amend new set of credentials to the passwords.txt file. Passwords.txt is located in the same folder with the program files for this assignment.

After the credentials have been validated, the menu.py is launched. Depending on the user’s choice, the functionality of EC2.py, EBSStorage.py, S3.py or monitor.py can be accessed via their objects.